feat: renamed test file to main, added graph and run
Also optimized environment to time out when mario stands still
