--- fitness_values.txt
@@ -1,8 +1,0 @@
-Generation: 0 - Best: 691.8200000000514 - Avg: 292.13993333333406 - Min: 139.27900000000167
-Generation: 1 - Best: 548.6000000000329 - Avg: 328.1407200000003 - Min: 139.27900000000167
-Generation: 2 - Best: 709.916000000049 - Avg: 334.6445200000005 - Min: 139.27900000000167
-Generation: 3 - Best: 985.8470000000035 - Avg: 353.09578000000147 - Min: 139.27900000000167
-Generation: 4 - Best: 985.8470000000035 - Avg: 379.66164000000504 - Min: 139.27900000000167
-Generation: 5 - Best: 999.7740000000052 - Avg: 429.2702866666741 - Min: 139.27900000000167
-Generation: 6 - Best: 1214.9670000000008 - Avg: 465.2484933333463 - Min: 139.27900000000167
-Generation: 7 - Best: 999.7740000000052 - Avg: 461.82494666667935 - Min: 139.27900000000167
